codes/requirements.txt
# Fundamentals
numpy
pyyaml
tb-nightly
future
scp
tqdm
matplotlib
scipy
munch
tqdm
scp
tensorboard
orjson
einops
lambda-networks
mup

# For image generation stuff
opencv-python
kornia
pytorch_ssim
gsa-pytorch
pytorch_fid==0.1.1

# For audio generation stuff
inflect==0.2.5
librosa==0.6.0
Unidecode==1.0.22
tgt == 1.4.4
pyworld == 0.2.10
audio2numpy
SoundFile

# For text stuff
transformers
tokenizers
jiwer  # calculating WER
omegaconf

# lucidrains stuff
vector_quantize_pytorch
linear_attention_transformer
rotary-embedding-torch
axial_positional_embedding
g-mlp-pytorch
x-clip
x_transformers

bitsandbytes
lion-pytorch==0.0.7
triton==2.0.0a2
